requirements.txt
Django==2.0.2
django-admin-list-filter-dropdown==1.0.1
django-import-export              1.0.1 
